Capabilities support Ticket #764
Ebook provides a way to determine the capabilities of the contact
store being connected to.  These capabilities are translated into
opensync capabilities and set during discover.

There is no way to ask ECal for capabilities so no capabilities are
added for the other objtypes.  I think this is the same as saying
all capabilities are supported which is the best we can do at the
moment.

git-svn-id: http://svn.opensync.org/plugins/evolution2@4413 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e